Tried implementing an incremental tryApply.
This is slower than the old style (probably due to creating and destroying a closure?), and it was a bit bug prone -- the scariest part was that replacements had to be (redundantly) checked for validity each time. This is probably an artifact of two things:

1. For multi-pattern rules, a match index found for an earlier pattern can be invalidated by replacements happening in later patterns; but there's no affordance now for resetting the counters of earlier patterns.
2. For patterns with ellipses, multiple candidates are generated and replacements for earlier ones can invalidate later ones.

It might be possible to fix (1) by checking, after doing a replacement, if the previous patterns are still satisfied (from the first down towards the second-to-last); if not, break up to that level in the loop.
Fixing (2) requires turning the ellipsis matching a bit inside out, so the application function is called once for each instantiation of the ellipsis.