NOT query operator.
Note that it still progresses the counter.
Uncomment the state store and restore terms to make it negative lookahead.
Does
`not (a then b then c)`
mean the same as:
`not a then not b then not c`?
No, because `a b d` is true for the former and false for the latter. It's more like:
`(not a) or (a then not b) or (a then b then not c)`.
But the challenging question is, how far does this progress the counter?: 0, 1, or 2 steps! It depends on how early the difference is detected! Well, that seems way wrong!